feat: Autenticação implementada nas rotas que precisam
Foi adicionada autenticação nas rotas nas quais se faz necessário seu uso. Para isso, algumas rotas precisaram passar por uma reformulação e/ou refatorações.

+++ users/routes/delete.go
@@ -2,6 +2,7 @@ package routes
 
 import (
 	"drink-counter-api/users/models"
+	UserServices "drink-counter-api/users/services"
 	DatabaseErrors "drink-counter-api/utils/db_errors"
 	SchemaErrors "drink-counter-api/utils/schema_errors"
 	"encoding/json"
@@ -15,6 +16,7 @@ import (
 // Deletes a user (token required). Only the user who created the user can delete it's own user.
 func DeleteUserHandler(db *gorm.DB, w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "application/json")
+
 	id, err := strconv.ParseUint(mux.Vars(r)["id"], 10, 32)
 	if err != nil {
 		w.WriteHeader(http.StatusBadRequest)
@@ -23,11 +25,17 @@ func DeleteUserHandler(db *gorm.DB, w http.ResponseWriter, r *http.Request) {
 		})
 		return
 	}
+
+	if UserServices.AuthService(w, r.Header.Get("Authorization"), uint(id)) == nil{
+		return
+	}
+
 	var user models.User
 	result := db.Delete(&user, "id = ?", id)
 	if DatabaseErrors.CheckDatabaseErrors(result.Error, w, "User") {
 		return
 	}
+
 	if result.RowsAffected == 0 {
 		w.WriteHeader(http.StatusNotFound)
 		json.NewEncoder(w).Encode(SchemaErrors.ErrorResponse{
